Refactored Feature: Authentication
