+++ lib/src/features/authentication/presentation/pages/login_page.dart
@@ -6,9 +6,9 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:the_apple_sign_in/the_apple_sign_in.dart';
 
-import '../Services/auth_service.dart';
-import '../components/custom_textfield.dart';
-import 'forgot_password_page.dart';
+import '../../../../../Services/auth_service.dart';
+import '../../../../../components/custom_textfield.dart';
+import '../../../../../pages/forgot_password_page.dart';
 
 class LoginPage extends StatefulWidget {
   const LoginPage({super.key});
